gameplay: fix enforcer "piercing" through damage vulnerability
add an exclusion for the mult_dmgtaken_from_bullets. the Pain Train
uses this to add bullet vulnerability, so the enforcer actually does less damage
in this case since it "pierces" the vulnerability. technically, the pierce resistance
attribute should also do this check for all other damage types for compatibility
with attribute mods, but that's a more extensive change than just addressing it
for the vanilla case.