fix: mitigate noisemaker spam
some cheaters use noisemakers to annoy other players in-game with
obnoxious sounds. add some integrity checks to sync up client and
server state properly.

first is a pretty reasonable change, check if the player
is alive and on the right team. this already happens on the
client, so this just verifies.

next is some adjustments for sound duration.
SV_GetSoundDuration actually isn't implemented for dedi
(maybe could be implemented now, since it's assumed
to be implemented on the server in some other cases too)
I did some minimal backports from CS:GO to get sound
duration access working in the case SV_GetSoundDuration
will be implemented in the future. CS:GO doesn't implement it
either. So, for now, I just extended to delay to 5 seconds, assuming
this bug is why the more popular duck journal has that delay.

Finally, we trust the client to tell the GC to consume a charge.
Obviously, it would be ideal to defer some sort of server consumption
to fix the original stated problem of wanting no audible delay on
noisemakers by contacting GC. So instead, let's just track the quantity
on the server to make sure there's no shenanigans going on.