gameplay: explosive headshot does bullet damage instance
Historically, explosive headshot was a damage over time applied
as bleed damage, but was then changed to deal a single damage
instance. Bleed ignores resistances, including not counting as ranged
damage for Fists of Steel heavies, which some people say makes
this mechanic problematic. However, this does change damage values
and effectiveness of Sniper in MvM so this may be ignored if not desired.

ref: ValveSoftware/Source-1-Games#4619