Add VScript convar allowlists
This PR adds a VScript convar allowlist for the HL2MP and TF2 mod, which is currently missing and requires the user to copy it over from TF2. The Valve titles on Steam, with the exception of TF2 also have the same issue, and therefore VScript cannot set any convars there.

In addition, the allowlist has been extended to have more harmless convars such as all the AI or NextBot ones. For TF2 the recently added convars such as tf_mvm_defenders_team_size or tf_gamemode_community are also now present.

This PR has two commits, one that contains the allowlist for HL2MP and TF2 mod, and one for the Valve titles. The commit for HL2MP and TF2 mod should be cherry picked, rather than merging this whole PR.