fix: crash on damage filter
sometimes, attackers are not defined.
this NULL check has precedent in other filters,
so we do it here too to make sure we don't crash
in the case of incorrect take damage info. can happen
with some custom stuff.

also, fix the known case of a null attacker for
the passtime ball.