fix: clean up sound duration usage
SoundDuration does not work on dedicated servers
For one, SV_GetSoundDuration is not implemented for
them. Two, sounds are not shipped on dedicated servers.

So, it's pointless and misleading to query the sound
duration on servers. Save a little bit by letting the sound
emit function skip trying to query for these sounds.

I replaced sound durations as needed with a sensible
sound duration for different cooldowns and such, so
sounds do not spam anymore, thinks run properly, etc.