make min viewmodels the default
this is an unfortunate change. the big viewmodels are artistically intended,
big and bulky weapons in your face give them weight and allow more
detail in the animations.

however, this very same artistic vision creates a feeling of clunkiness and
claustrophobia in players, especially ones who are much more accustomed
to newer FPS titles.

therefore, we are making min viewmodels the default